refactor(datasource-api): split contracts by responsibility
