+++ yak-ops-plugins/yak-ops-plugin-datasource/yak-ops-plugin-datasource-api/src/main/java/io/yak/ops/spi/datasource/catalog/package-info.java
@@ -1,0 +1,2 @@
+/** Catalog 查询条件与资源定位模型。 */
+package io.yak.ops.spi.datasource.catalog;
