+++ yak-ops-plugins/yak-ops-plugin-datasource/yak-ops-plugin-datasource-api/src/main/java/io/yak/ops/spi/datasource/metadata/package-info.java
@@ -1,0 +1,2 @@
+/** 数据源表、字段等结构化元数据模型。 */
+package io.yak.ops.spi.datasource.metadata;
