+++ yak-ops-plugins/yak-ops-plugin-datasource/yak-ops-plugin-datasource-api/src/main/java/io/yak/ops/spi/datasource/package-info.java
@@ -1,2 +1,6 @@
-/** 数据源插件、参数规范化和 Catalog 元数据扩展契约。 */
+/**
+ * 数据源插件稳定 SPI。
+ *
+ * <p>插件、Catalog、连接和异常等核心契约保留在根包；Catalog 请求、元数据和查询结果模型分别位于对应子包。
+ */
 package io.yak.ops.spi.datasource;
