+++ yak-ops-plugins/yak-ops-plugin-datasource/yak-ops-plugin-datasource-api/src/main/java/io/yak/ops/spi/datasource/query/package-info.java
@@ -1,0 +1,2 @@
+/** 数据预览和轻量查询结果模型。 */
+package io.yak.ops.spi.datasource.query;
