+++ yak-ops-plugins/yak-ops-plugin-datasource/yak-ops-plugin-datasource-jdbc/src/main/java/io/yak/ops/plugin/database/jdbc/GenericJdbcCatalog.java
@@ -2,14 +2,14 @@ package io.yak.ops.plugin.database.jdbc;
 
 import io.yak.ops.common.enums.datasource.DataSourceDbType;
 import io.yak.ops.spi.datasource.DataSourceCatalog;
-import io.yak.ops.spi.datasource.DataSourceCatalogQuery;
-import io.yak.ops.spi.datasource.DataSourceColumn;
 import io.yak.ops.spi.datasource.DataSourcePluginException;
 import io.yak.ops.spi.datasource.DataSourcePluginException.Operation;
-import io.yak.ops.spi.datasource.DataSourceQueryColumn;
-import io.yak.ops.spi.datasource.DataSourceQueryResult;
-import io.yak.ops.spi.datasource.DataSourceTable;
-import io.yak.ops.spi.datasource.DataSourceTablePath;
+import io.yak.ops.spi.datasource.catalog.DataSourceCatalogQuery;
+import io.yak.ops.spi.datasource.catalog.DataSourceTablePath;
+import io.yak.ops.spi.datasource.metadata.DataSourceColumn;
+import io.yak.ops.spi.datasource.metadata.DataSourceTable;
+import io.yak.ops.spi.datasource.query.DataSourceQueryColumn;
+import io.yak.ops.spi.datasource.query.DataSourceQueryResult;
 import java.sql.Connection;
 import java.sql.DatabaseMetaData;
 import java.sql.DriverManager;
